feat: flash services, code cleanup, common logic refactor

+++ modules/ice_services/include/ice_srv_ping.h
@@ -4,6 +4,6 @@
 #include <stdint.h>
 #include "ice_boot_protocol.h"
 
-ice_service_status_t ice_srv_ping(uint8_t *resp_buf, uint16_t *resp_len);
+ice_service_status_t ice_srv_ping(uint8_t *resp_buf, uint32_t *resp_len);
 
 #endif // ICE_SRV_PING_H
